+++ src/SoundDeck.PI/js/components/app.jsx
@@ -1,13 +1,16 @@
 import React from "react";
+import AppMultimediaControlsSettings from "./appMultimediaControlsSettings";
 import CaptureAudioSettings from "./captureAudioSettings";
 import PlayAudioSettings from "./playAudioSettings";
 import SamplerSettings from "./samplerSettings";
-import SessionMultimediaControlsSettings from "./sessionMultimediaControlsSettings";
 import SetAppAudioDeviceSettings from "./setAppAudioDeviceSettings"
 
 class App extends React.Component {
     render() {
         switch (this.props.uuid) {
+            case "com.geekyeggo.sounddeck.appmultimediacontrols":
+                return <AppMultimediaControlsSettings />;
+
             case "com.geekyeggo.sounddeck.clipaudio":
                 return <CaptureAudioSettings showDuration={true} />;
 
@@ -20,14 +23,11 @@ class App extends React.Component {
             case "com.geekyeggo.sounddeck.sampler":
                 return <SamplerSettings />;
 
-            case "com.geekyeggo.sounddeck.sessionmultimediacontrols":
-                return <SessionMultimediaControlsSettings />;
-
             case "com.geekyeggo.sounddeck.setappaudiodevice":
-                return <SetAppAudioDeviceSettings />
+                return <SetAppAudioDeviceSettings />;
 
             default:
-                return <div />
+                return <div />;
         }
     }
 }
