+++ src/SoundDeck.Plugin/Images/AppMultimediaControls/Action.png
