+++ src/SoundDeck.Plugin/Images/AppMultimediaControls/Action@2x.png
