+++ src/SoundDeck.Plugin/Images/AppMultimediaControls/Key.png
