+++ src/SoundDeck.Plugin/Images/AppMultimediaControls/Key@2x.png
